agents/researcher-agent/pyproject.toml
[project]
name = "researcher-agent"
version = "0.1.0"
description = "LangGraph Researcher Agent"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "langgraph>=0.2.0",
  "ddgs>=9.0.0",
  "azure-identity>=1.17.0",
  "langchain-openai>=0.2.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
addopts = "-q"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
